+++ include/BossConversation.h
@@ -64,11 +64,11 @@ class BossConversation : public KeyMapAct
         void OnHide();
         void OnNext();
         void OnEvent(int,Key,bool);
-        void LoadConversation(const std::string& s,const std::string& basePath,Bgm* bgm,double bgmBpm);
+        void LoadConversation(const std::string& s,const std::string& basePath,Bgm* bgm,float bgmBpm);
     protected:
     private:
         Bgm* m_bgm;
-        double m_bgmBpm;
+        float m_bgmBpm;
 
         ACGCross::Galgame::TextBox* m_text;
         std::string m_basePath;
