+++ include/CSVWRollBgp.h
@@ -11,9 +11,9 @@ class CSVWRollBgp : public Background
         struct SingleBG{
             SDL_Texture* tex;
             int w,h;
-            double depth;double depthTrg;int depthLostTime = 0;
-            double r1x,r2x;
-            Uint8 alpha;double alphaTrg;int alphaLostTime = 0;
+            double depth;float depthTrg;int depthLostTime = 0;
+            float r1x,r2x;
+            Uint8 alpha;float alphaTrg;int alphaLostTime = 0;
             int y;int yTrg;int yLostTime = 0;
         };
         std::vector<SingleBG> m_bgs;
@@ -32,13 +32,13 @@ class CSVWRollBgp : public Background
         std::queue <SingleCmd> m_cmds;
         int m_cnt;
 
-        double m_spd;double m_spdTrg;int m_spdLostTime = 0;
+        float m_spd;float m_spdTrg;int m_spdLostTime = 0;
     public:
         void LoadCSV(const std::string& stage,const std::string& bgcsv);
         void OnNext();
         void OnDraw();
         void Clear();
-        inline void SetSpeed(double spd){m_spd = spd;}
+        inline void SetSpeed(float spd){m_spd = spd;}
 };
 
 extern CSVWRollBgp background;
