+++ include/ItemMgr.h
@@ -16,7 +16,7 @@ class ItemMgr
         static void Init();
         void OnDraw();
         void OnNext();
-        void AddItem(ItemType i,double boom,int x,int y,int num);  //类型,爆炸幅度,x,y,个数
+        void AddItem(ItemType i,float boom,int x,int y,int num);  //类型,爆炸幅度,x,y,个数
         void Clear();
         void GetAll(int playerNum);
         void GetItem(int num,int playerNum);
@@ -27,15 +27,15 @@ class ItemMgr
         struct Item{
             bool live = false;
             ItemType type;
-            double x,y;
+            float x,y;
             enum{
                 DOWNING,
                 GETTING
             }state;
             int boom;
             float gettingSpd = 0;
-            double ySpd = 0;
-            double boomSpdX = 0;
+            float ySpd = 0;
+            float boomSpdX = 0;
 
             int gettingToPlayer;
         }m_items[750];
