+++ include/PlayerBullet.h
@@ -3,12 +3,12 @@
 
 class PlayerBullet{
     private:
-	inline void setBullet(int num,double x,double y,double power,int style);
+	inline void setBullet(int num,float x,float y,float power,int style);
 
 	int m_searchTop = 0;
 	struct PlayerBulletObj{
 		int style=-1;    //-1 == DEATH
-		double x,y,pow;
+		float x,y,pow;
 	}m_bullets[512];
 
 	static struct PlayerBulletStyle{
@@ -18,12 +18,12 @@ class PlayerBullet{
 
 	public:
     static void Init();
-	void Add(double x,double y,double power,int style);
+	void Add(float x,float y,float power,int style);
 	void OnNext();
 	void OnDraw();
 	void Clear();
 	void Kill(int num);
-	inline double GetPower(int num){
+	inline float GetPower(int num){
         return m_bullets[num].pow;
 	}
