+++ include/ShotParttenAPI.h
@@ -2,7 +2,7 @@
 #include "BulletMgr.h"
 
 /* 仅对Boss符卡有效的API */
-typedef void(*SCPartten)(Boss* b,int cnt,int scnt,int& image,double& boss_x,double& boss_y,double& boss_spd,double& boss_aspd,double& boss_angle,double hp,const std::vector<int>& bullets,Snow::Bundle<256>& data);
+typedef void(*SCPartten)(Boss* b,int cnt,int scnt,int& image,float& boss_x,float& boss_y,float& boss_spd,float& boss_aspd,float& boss_angle,float hp,const std::vector<int>& bullets,Snow::Bundle<256>& data);
 //Boss 指针，Boss总帧数，符卡生存帧数，使用中的Boss图像，Boss坐标x，Boss坐标y，Boss速度，Boss加速度，Boss移动角度，符卡生命剩余，子弹表
 extern SCPartten scPartten [];
 
@@ -16,7 +16,7 @@ extern SCPartten scPartten [];
 #define BltKill(b) bulletMgr.Kill(b)
 #include "Snow/Debug.h"
 #include "Boss.h"
-inline int SCCreateBlt(Boss* s,double x,double y,double angle,double spd,int minLiveTime,int style)
+inline int SCCreateBlt(Boss* s,float x,float y,float angle,float spd,int minLiveTime,int style)
 {
     int n = bulletMgr.Alloc(x,y,style);
     if(n == -1) return -1;
